fix: warn user to quit firefox before root install
At least on Linux, potentially other OSes as well, Firefox appears to
overwrite the cert8.db on exit, apparently from an in-memory copy loaded
at startup. So if the user has firefox running, and we install the root
cert, and they quit firefox, they'll overwrite the cert8.db and lose the
root cert. If firefox is not running at install time, then we'll be
fine.